requirements.txt
tensorflow
numpy
pillow
nltk
flask
tqdm
matplotlib
